# openapi.yaml
openapi: 3.1.0
info:
  title: SKINgenius API
  version: 1.0.0
  description: |
    Free, on-device AI skin analysis. No per-scan costs. Privacy-first.

    Scan images are analyzed on the user's device (MediaPipe + Gemma 4).
    Only results are sent to this API for persistence and recommendations.

    ## Authentication
    Most endpoints require authentication via Supabase session cookie.
    The API checks for `sb-access-token` cookie or `x-test-user-id` header (dev only).

    ## Base URL
    Production: `https://skingenius-sigma.vercel.app`
  contact:
    name: SKINgenius
    email: jasonopland@msn.com
  license:
    name: Proprietary
servers:
  - url: https://skingenius-sigma.vercel.app
    description: Production
paths:
  /api/v1/scan:
    post:
      operationId: createScan
      summary: Submit scan results
      description: |
        Accepts on-device analysis results (preferred) OR a raw base64 image (legacy).
        The scan pipeline runs entirely on the user's phone when `analysis_results` is provided.
        This endpoint persists results to the `scan_results` table and returns the saved record.
      tags:
        - Scan
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/V1ScanRequestOnDevice'
                - $ref: '#/components/schemas/V1ScanRequestLegacy'
            examples:
              on-device:
                summary: On-device analysis results
                value:
                  analysis_results:
                    scan_id: null
                    timestamp: "2026-05-23T12:00:00Z"
                    quality_assessment:
                      is_valid_format: true
                      width: 1280
                      height: 1280
                      aspect_ratio: 1.0
                      file_size_bytes: 245760
                      face_detected: true
                      blur_score: 8.5
                      lighting_score: 7.2
                    conditions:
                      - condition_id: "acne-vulgaris"
                        name: "Acne Vulgaris"
                        confidence: 0.92
                        severity: "moderate"
                        features: ["papules", "pustules"]
                        zone: "t-zone"
                    skin_zones:
                      - zone: "t-zone"
                        primary_concern: "Acne Vulgaris"
                        description: "Moderate comedonal and inflammatory acne"
                        severity: "moderate"
                        confidence: 0.92
                    metadata:
                      capture_method: "camera"
                      skin_tone: 3
                      processed: true
                      model_version: "v1-ondevice"
                  capture_method: "camera"
                  skin_tone: 3
              legacy:
                summary: Legacy raw image upload
                value:
                  image: "/9j/4AAQ..."
                  capture_method: "camera"
                  skin_tone: 3
      responses:
        '200':
          description: Scan results persisted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1ScanResponse'
              example:
                data:
                  scan_id: "550e8400-e29b-41d4-a716-446655440000"
                  timestamp: "2026-05-23T12:00:00Z"
                  quality_assessment:
                    is_valid_format: true
                    width: 1280
                    height: 1280
                    face_detected: true
                    blur_score: 8.5
                    lighting_score: 7.2
                  conditions:
                    - condition_id: "acne-vulgaris"
                      name: "Acne Vulgaris"
                      confidence: 0.92
                      severity: "moderate"
                      features: ["papules", "pustules"]
                      zone: "t-zone"
                  skin_zones:
                    - zone: "t-zone"
                      primary_concern: "Acne Vulgaris"
                      description: "Moderate acne"
                      severity: "moderate"
                      confidence: 0.92
                  metadata:
                    capture_method: "camera"
                    skin_tone: 3
                    processed: true
                    model_version: "v1-ondevice"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Missing or invalid field: capture_method (camera | gallery)"
        '401':
          description: Unauthorized — no valid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
        '422':
          description: Image failed quality gate (legacy mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1ScanResponse'
              example:
                error: "Image failed quality gate"
                data:
                  scan_id: null
                  timestamp: "2026-05-23T12:00:00Z"
                  quality_assessment:
                    is_valid_format: true
                    face_detected: false
                  conditions: []
                  skin_zones: []
                  metadata:
                    capture_method: "camera"
                    skin_tone: 3
                    processed: false
                    model_version: "v1-mock"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie. Required for authenticated endpoints.
  schemas:
...
    V1ScanRequestOnDevice:
      type: object
      required: [analysis_results, capture_method, skin_tone]
      properties:
        analysis_results:
          $ref: '#/components/schemas/V1ScanResponseData'
        capture_method:
          type: string
          enum: [camera, gallery]
        skin_tone:
          type: integer
          minimum: 1
          maximum: 6
    V1ScanRequestLegacy:
      type: object
      required: [image, capture_method, skin_tone]
      properties:
        image:
          type: string
          description: Base64-encoded JPEG/PNG image
        capture_method:
          type: string
          enum: [camera, gallery]
        skin_tone:
          type: integer
          minimum: 1
          maximum: 6
    V1ScanResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/V1ScanResponseData'
        error:
          type: string
        detail:
          type: string
    V1ScanResponseData:
      type: object
      required: [scan_id, timestamp, quality_assessment, conditions, skin_zones, metadata]
      properties:
        scan_id:
          type: string
          format: uuid
          nullable: true
        timestamp:
          type: string
          format: date-time
        quality_assessment:
          $ref: '#/components/schemas/V1QualityAssessment'
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/V1DetectedCondition'
        skin_zones:
          type: array
          items:
            $ref: '#/components/schemas/V1SkinZone'
        metadata:
          $ref: '#/components/schemas/V1ScanMetadata'
    V1DetectedCondition:
      type: object
      required: [condition_id, name, confidence, severity, features, zone]
      properties:
        condition_id:
          type: string
        name:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        severity:
          type: string
          enum: [mild, moderate, severe]
        features:
          type: array
          items:
            type: string
        zone:
          type: string
    V1SkinZone:
      type: object
      required: [zone, primary_concern, description, severity, confidence]
      properties:
        zone:
          type: string
        primary_concern:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [mild, moderate, severe]
        confidence:
          type: number
          minimum: 0
          maximum: 1
    V1QualityAssessment:
      type: object
      required: [is_valid_format, file_size_bytes, face_detected]
      properties:
        is_valid_format:
          type: boolean
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        aspect_ratio:
          type: number
          nullable: true
        file_size_bytes:
          type: integer
        face_detected:
          type: boolean
        blur_score:
          type: number
          nullable: true
        lighting_score:
          type: number
          nullable: true
    V1ScanMetadata:
      type: object
      required: [capture_method, skin_tone, processed, model_version]
      properties:
        capture_method:
          type: string
        skin_tone:
          type: integer
        processed:
          type: boolean
        model_version:
          type: string
        warnings:
          type: array
          items:
            type: string
        storage_failed:
          type: boolean
        storage_error:
          type: string
        auto_detected_skin_tone:
          type: object
          properties:
            type:
              type: integer
            confidence:
              type: number

    # ─── Recommendation Schemas ──────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        detail:
          type: string
        path:
          type: string